Rename print file sorting module in line with our standard
